Hierarchy/public/requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
pillow==10.3.0
